+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Features
 
+-   [`34ed9af`](https://github.com/stdlib-js/stdlib/commit/34ed9af63e302a5f4a759c7ef9b4f048b3ef89d4) - add `ndarray/base/unary-reduce-strided1d-dispatch-by`
 -   [`2460ce8`](https://github.com/stdlib-js/stdlib/commit/2460ce82935de6512ee0979cc3b3918924af5ae1) - add `unaryReduceStrided1dBy` to namespace
 -   [`b228b5e`](https://github.com/stdlib-js/stdlib/commit/b228b5e89da183b38160c6cd6f9908a07918dd09) - add `ndarray/base/unary-reduce-strided1d-by` [(#7214)](https://github.com/stdlib-js/stdlib/pull/7214)
 -   [`67e9602`](https://github.com/stdlib-js/stdlib/commit/67e9602131f6e714d98de70ce6809cd9d327b02f) - add `someBy` to namespace
@@ -307,6 +308,8 @@
 
 ### Bug Fixes
 
+-   [`c86cb96`](https://github.com/stdlib-js/stdlib/commit/c86cb96872dce4d7b508daf43886e10c0fedad8f) - update require path
+-   [`0ddb52b`](https://github.com/stdlib-js/stdlib/commit/0ddb52b2f0719c4dce9aa2e5a7fd1d1112dae6e5) - update require path
 -   [`3f6c0ff`](https://github.com/stdlib-js/stdlib/commit/3f6c0ffd9568665b54d0547b0550b2c96505bed6) - address increment bug
 -   [`4d29349`](https://github.com/stdlib-js/stdlib/commit/4d29349017e6d62d7f72aa0c3a86a24fdd83e86a) - address increment bug
 -   [`d1f5820`](https://github.com/stdlib-js/stdlib/commit/d1f58208fc95174fcfce03f00df278a8190578de) - address increment bug
@@ -434,6 +437,11 @@ A total of 17 issues were closed in this release:
 
 <details>
 
+-   [`34ed9af`](https://github.com/stdlib-js/stdlib/commit/34ed9af63e302a5f4a759c7ef9b4f048b3ef89d4) - **feat:** add `ndarray/base/unary-reduce-strided1d-dispatch-by` _(by Athan Reines)_
+-   [`970b3be`](https://github.com/stdlib-js/stdlib/commit/970b3be65917ec995e5d8c0e06fbebd5870a4687) - **docs:** fix description _(by Athan Reines)_
+-   [`c86cb96`](https://github.com/stdlib-js/stdlib/commit/c86cb96872dce4d7b508daf43886e10c0fedad8f) - **fix:** update require path _(by Athan Reines)_
+-   [`0ddb52b`](https://github.com/stdlib-js/stdlib/commit/0ddb52b2f0719c4dce9aa2e5a7fd1d1112dae6e5) - **fix:** update require path _(by Athan Reines)_
+-   [`dded8bb`](https://github.com/stdlib-js/stdlib/commit/dded8bb627d2634c2e244a9a74bc320511d66263) - **docs:** fix description _(by Athan Reines)_
 -   [`03e5c54`](https://github.com/stdlib-js/stdlib/commit/03e5c54b90e538bee276484281d2e4553d7ec83e) - **docs:** update namespace table of contents [(#7233)](https://github.com/stdlib-js/stdlib/pull/7233) _(by stdlib-bot)_
 -   [`2460ce8`](https://github.com/stdlib-js/stdlib/commit/2460ce82935de6512ee0979cc3b3918924af5ae1) - **feat:** add `unaryReduceStrided1dBy` to namespace _(by Athan Reines)_
 -   [`b228b5e`](https://github.com/stdlib-js/stdlib/commit/b228b5e89da183b38160c6cd6f9908a07918dd09) - **feat:** add `ndarray/base/unary-reduce-strided1d-by` [(#7214)](https://github.com/stdlib-js/stdlib/pull/7214) _(by Muhammad Haris, Athan Reines)_